pemesanan sudah bisa

+++ controllers/pemesanan.controller.js
@@ -18,13 +18,7 @@ exports.addPemesanan = async (request, response) => {
     where: {
       [Op.and]: [{ nomor_kamar: { [Op.substring]: nomor_kamar } }],
     },
-    attributes: [
-      "id",
-      "nomor_kamar",
-      "tipeKamarId",
-      "createdAt",
-      "updatedAt",
-    ],
+    attributes: ["id", "nomor_kamar", "tipeKamarId", "createdAt", "updatedAt"],
   });
 
   let nama_user = request.body.nama_user;
@@ -53,36 +47,43 @@ exports.addPemesanan = async (request, response) => {
       tgl_check_in: request.body.check_in,
       tgl_check_out: request.body.check_out,
       nama_tamu: request.body.nama_tamu,
-      jumlah_kamar: request.body.jumlah_kamar,
+      jumlah_kamar: 1, // Hanya satu jumlah kamar yang diizinkan
       tipeKamarId: room.tipeKamarId,
       status_pemesanan: request.body.status,
       userId: userId.id,
     };
     console.log(newData);
     let roomCheck = await sequelize.query(
-      `SELECT * FROM detail_pemesanans WHERE kamarId = ${room.id} AND tgl_akses= "${request.body.check_in}" ;`
+      `SELECT * FROM detail_pemesanans WHERE kamarId = ${room.id} AND tgl_akses >= "${request.body.check_in}" AND tgl_akses <= "${request.body.check_out}" ;`
     );
 
     if (roomCheck[0].length === 0) {
+      const tglCheckIn = new Date(request.body.check_in);
+      const tglCheckOut = new Date(request.body.check_out);
+      const diffTime = Math.abs(tglCheckOut - tglCheckIn);
+      const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+
       pemesananModel
         .create(newData)
         .then((result) => {
           let pemesananID = result.id;
           let detailsOfPemesanan = request.body.details_of_pemesanan;
+          let detailData = [];
 
-          for (let i = 0; i < detailsOfPemesanan.length; i++) {
-            detailsOfPemesanan[i].pemesananId = pemesananID;
+          for (let i = 0; i < diffDays; i++) {
+            let newDetail = {
+              pemesananId: pemesananID,
+              kamarId: room.id,
+              tgl_akses: new Date(
+                tglCheckIn.getTime() + i * 24 * 60 * 60 * 1000
+              ),
+              harga: detailsOfPemesanan[0].harga,
+            };
+            detailData.push(newDetail);
           }
 
-          let newDetail = {
-            id: pemesananID,
-            kamarId: room.id,
-            tgl_akses: result.tgl_check_in,
-            harga: detailsOfPemesanan[0].harga,
-          };
-
           detailsOfPemesananModel
-            .create(newDetail)
+            .bulkCreate(detailData)
             .then((result) => {
               return response.json({
                 success: true,
@@ -118,13 +119,7 @@ exports.updatePemesanan = async (request, response) => {
     where: {
       [Op.and]: [{ nomor_kamar: { [Op.substring]: nomor_kamar } }],
     },
-    attributes: [
-      "id",
-      "nomor_kamar",
-      "tipeKamarId",
-      "createdAt",
-      "updatedAt",
-    ],
+    attributes: ["id", "nomor_kamar", "tipeKamarId", "createdAt", "updatedAt"],
   });
 
   let nama_user = request.body.nama_user;
@@ -136,13 +131,13 @@ exports.updatePemesanan = async (request, response) => {
 
   let newData = {
     nomor_pemesanan: request.body.nomor_pemesanan,
-    nama_pemesanan: request.body.nama_pemesanan,
-    email_pemesanan: request.body.email_pemesanan,
-    tgl_pemesanan: request.body.tanggal_pemesanan,
+    nama_pemesan: request.body.nama_pemesan,
+    email_pemesan: request.body.email_pemesan,
+    tgl_pemesanan: request.body.tgl_pemesanan,
     tgl_check_in: request.body.check_in,
     tgl_check_out: request.body.check_out,
     nama_tamu: request.body.nama_tamu,
-    jumlah_kamar: request.body.jumlah_kamar,
+    jumlah_kamar: 1, // Hanya satu jumlah kamar yang diizinkan
     tipeKamarId: room.tipeKamarId,
     status_pemesanan: request.body.status,
     userId: userId.id,
@@ -150,47 +145,62 @@ exports.updatePemesanan = async (request, response) => {
 
   let pemesananID = request.params.id;
 
-  pemesananModel
-    .update(newData, { where: { id: pemesananID } })
-    .then(async (result) => {
-      await detailsOfPemesananModel.destroy({
-        where: { id: pemesananID },
-      });
+  try {
+    const existingPemesanan = await pemesananModel.findByPk(pemesananID);
 
-      let detailsOfPemesanan = request.body.details_of_pemesanan;
-
-      for (let i = 0; i < detailsOfPemesanan.length; i++) {
-        detailsOfPemesanan[i].id = pemesananID;
-      }
-
-      let newDetail = {
-        id: pemesananID,
-        kamarId: room.id,
-        tgl_akses: detailsOfPemesanan[0].tgl_akses,
-        harga: detailsOfPemesanan[0].harga,
-      };
-
-      detailsOfPemesananModel
-        .create(newDetail)
-        .then((result) => {
-          return response.json({
-            success: true,
-            message: ` transaction has been update`,
-          });
-        })
-        .catch((error) => {
-          return response.json({
-            success: false,
-            message: error.message,
-          });
-        });
-    })
-    .catch((error) => {
+    if (!existingPemesanan) {
       return response.json({
         success: false,
-        message: error.message,
+        message: `Pemesanan dengan ID ${pemesananID} tidak ditemukan`,
       });
+    }
+
+    let roomCheck = await sequelize.query(
+      `SELECT * FROM detail_pemesanans WHERE kamarId = ${room.id} AND tgl_akses >= "${request.body.check_in}" AND tgl_akses <= "${request.body.check_out}" ;`
+    );
+
+    if (roomCheck[0].length === 0) {
+      const tglCheckIn = new Date(request.body.check_in);
+      const tglCheckOut = new Date(request.body.check_out);
+      const diffTime = Math.abs(tglCheckOut - tglCheckIn);
+      const diffDays = Math.ceil(diffTime / (1000 * 60 * 60 * 24));
+
+      await existingPemesanan.update(newData);
+
+      await detailsOfPemesananModel.destroy({
+        where: { pemesananId: pemesananID },
+      });
+
+      let detailsOfPemesanan = [];
+
+      for (let i = 0; i < diffDays; i++) {
+        let newDetail = {
+          pemesananId: pemesananID,
+          kamarId: room.id,
+          tgl_akses: new Date(tglCheckIn.getTime() + i * 24 * 60 * 60 * 1000),
+          harga: request.body.harga, // Ubah ini dengan logika perhitungan harga yang sesuai
+        };
+        detailsOfPemesanan.push(newDetail);
+      }
+
+      await detailsOfPemesananModel.bulkCreate(detailsOfPemesanan);
+
+      return response.json({
+        success: true,
+        message: `Pemesanan dengan ID ${pemesananID} berhasil diperbarui`,
+      });
+    } else {
+      return response.json({
+        success: false,
+        message: `Kamar yang anda pesanan sudah di booking`,
+      });
+    }
+  } catch (error) {
+    return response.json({
+      success: false,
+      message: error.message,
     });
+  }
 };
 
 //delete data
@@ -225,9 +235,9 @@ exports.deletePemesanan = async (request, response) => {
     });
 };
 
-//mendapatkan semua data 
+//mendapatkan semua data
 exports.getAllPemesanan = async (request, response) => {
-  const result = await pemesananModel.findAll()
+  const result = await pemesananModel.findAll();
 
   response.json({
     success: true,
@@ -236,17 +246,14 @@ exports.getAllPemesanan = async (request, response) => {
   });
 };
 
-//mendapatkan salah satu data 
+//mendapatkan salah satu data
 exports.find = async (request, response) => {
   let status = request.body.status;
 
   const result = await pemesananModel.findAll({
-    where:{
-
-      [Op.and]: [
-        { status_pemesanan: status},
-      ],
-    }
+    where: {
+      [Op.and]: [{ status_pemesanan: status }],
+    },
   });
 
   return response.json({
